+++ sdk/resourcemanager/virtualmachineimagebuilder/armvirtualmachineimagebuilder/fake/triggers_server.go
@@ -29,7 +29,7 @@ type TriggersServer struct {
 	BeginCreateOrUpdate func(ctx context.Context, resourceGroupName string, imageTemplateName string, triggerName string, parameters armvirtualmachineimagebuilder.Trigger, options *armvirtualmachineimagebuilder.TriggersClientBeginCreateOrUpdateOptions) (resp azfake.PollerResponder[armvirtualmachineimagebuilder.TriggersClientCreateOrUpdateResponse], errResp azfake.ErrorResponder)
 
 	// BeginDelete is the fake for method TriggersClient.BeginDelete
-	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted, http.StatusNoContent
+	// HTTP status codes to indicate success: http.StatusAccepted, http.StatusNoContent
 	BeginDelete func(ctx context.Context, resourceGroupName string, imageTemplateName string, triggerName string, options *armvirtualmachineimagebuilder.TriggersClientBeginDeleteOptions) (resp azfake.PollerResponder[armvirtualmachineimagebuilder.TriggersClientDeleteResponse], errResp azfake.ErrorResponder)
 
 	// Get is the fake for method TriggersClient.Get
@@ -182,9 +182,9 @@ func (t *TriggersServerTransport) dispatchBeginDelete(req *http.Request) (*http.
 		return nil, err
 	}
 
-	if !contains([]int{http.StatusOK, http.StatusAccepted, http.StatusNoContent}, resp.StatusCode) {
+	if !contains([]int{http.StatusAccepted, http.StatusNoContent}, resp.StatusCode) {
 		t.beginDelete.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusAccepted, http.StatusNoContent", resp.StatusCode)}
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusAccepted, http.StatusNoContent", resp.StatusCode)}
 	}
 	if !server.PollerResponderMore(beginDelete) {
 		t.beginDelete.remove(req)